backend/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.23.2
celery==5.3.4
python-dotenv==1.0.0
pymongo==4.5.0